PatientChildController and PatientChildrenController.
All of the patient child controllers were already inheriting from
PatientChildController, but it was only useful for mulit-associations,
single associations still needed to be implemented manually. I finally
generalized out the single association logic into
PatientChildController, and moved the multi-association logic to
the newly created PatientChildrenController.

As a result, we can remove most of the logic from all patient child
controllers. The notable exception is vital_signs, because it uses the
templates from another controller.