Global test type callbacks.
The amount of application-specific business logic in the test type
implementation was making me uncomfortable, so I implemented global
callbacks and moved this code into a global assign callback. Also
added some comments documentating the test type configuration, this
should make it clearer what is going on in the callbacks.