Enable any name to be used as a test type callback.
Due to the way test type registration was implemented, some callback
names (e.g., inspect) would fail silently because the conflict with
built-in methods. This is no longer the case. It's still possible for
a name to conflict if it's a Ruby reserved word, but there in that
case you can use self.whatever as a workaround.

 * lib/test_type.rb
   - TestType::Reservation now a subclass of ActiveSupport::BasicObject
 * lib/test_types.rb
   - rename results pix feed operation to inspect
 * app/views/vendor_test_plans/_pix_feed_action_area.html.erb
   - change results to inspect