Eliminate dependency on andand gem.
The andand idiom was always subject of debate in the Ruby community, and
in the end both Rails and Ruby 1.9 adopted a different idiom, the try
method. I prefer try for two reasons: It has the exact same semantics as
Object#send, which is a very common and well-known standard method whose
behavior won't change. Also, andand pretty much lost to try once Ruby
1.9 and then Rails got on the train.

This also gets most of the tests passing in JRuby 1.3.1.