SortOrder cleanup.
For some reason two of the sort order tests are failing under JRuby
1.3.1. I'm not sure of the cause, but it appears that controllers will
not raise UnknownAction for methods hidden with hide_action. My best
guess as to the cause is the latest JRuby Rack adapter maybe, but
really I still have no idea. I cleaned up the code in lib/sort_order.rb
while I was trying to get this working, so I'm checking it in even
though it didn't fix the two broken tests.

All the tests should still pass with JRuby 1.1.6.