Allow specification of test setup using callbacks.
There is a new callback available for test types. The setup callback can
be used to provide a starting point for gathering information for a
test, and is generally used to setup and render a form.

Both XDS tests are now set up using this facility, resulting in the
removal of two confusingly-named actions from xds_patients: query and
provide_and_register_setup.