Started new TestPlan base class for defining test types.
With major UX changes approaching for Laika, and with most of the
test-specific code now moved into test-type-specific callbacks, there is
an opportunity to rethink the test plan type. It's likely that the
existing test type system *and* the Kind and VendorTestPlan types will
all eventually be replaced by this facility.

A major advantage to using classes to define test types is documentation
and lack of duplicated effort. Current the YARD doc too can't really
see into test type definitions in a useful way. Moving to a class-based
solution will address this. Also by using classes to define test types I
can take advantage of Ruby features like inheritance and roles that
otherwise need to be implemented manually (as they are currently).