Replace some VendorTestPlan references with TestPlan.
I'm going about the removal of VendorTestPlan in a pretty methodical
way, by working in small commits and keeping the tests passing. In this
commit I'm replacing references to VendorTestPlan only in code that is
staying in Laika. I didn't bother to update code that's going to be
removed, mostly to allow me to work in smaller increments.

 * app/controllers/account_controller.rb,
   app/controllers/application_controller.rb
   - replace url helpers
 * app/controllers/users_controller.rb
   - index: do not display unclaimed vendors
 * app/models/user.rb
   - new method: count_test_plans
   - destroy dependent test plans
 * app/models/vendor.rb
   - remove method: unclaimed
   - update editable_by? to ignore admins
 * app/views/account/signup.rhtml
   - replace url helpers
 * app/views/users/index.html.erb
   - do not show count of test users when removing vendors, no longer used
 * config/routes.rb
   - remove unused xds routes
   - point root at TestPlansController#index