Random patient: spec and cleanups
I added a spec to test for simple failure to generate a random patient,
and ended up creating a standard way for Laika to efficiently select a
random record. In many cases we loaded every record into memory, sorted
them randomly, then picked the first one. This approach is aggressive in
its inefficiency. The new randomization API is used like this:

    class MyModel < ActiveRecord::Base
      extend RandomFinder
    end
    MyModel.find :random

MyModel.find(:random) counts the number of records available and picks a
single record at a randomized offset in that range. We were doing this
sometimes (e.g. Address#randomize) but not consistently.

I also made it possible to generate a patient even when there is no
lookup data present. It doesn't do anything magic: with no lookup data
the resulting patient is mostly empty. Beats having it blow up, though.