Removed unnecessary requires.
There was a lot of code that was needlessly requiring libraries that
were already loaded, or that were requiring libs by explictly putting
"lib/" at the start of the require path. It's important to require files
in a consistent way to avoid loading code multiple times and to permit
auto-reloading of code during development.

I removed or fixed all requires starting with "lib/", and removed
numerous instances where validation code was being required
unnecessarily.